requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.13.1
mysqlclient==2.2.7
pillow==11.1.0
djangorestframework-simplejwt==5.5.0
